履歴のレイアウトがずれていたので修正
