+++ androidApp/src/main/res/values/strings.xml
@@ -46,6 +46,7 @@
     <string name="point_use_confirm_execute_button">ポイントを使用する</string>
     <string name="point_history_acquire">ポイント獲得</string>
     <string name="point_history_use">ポイント使用</string>
+    <string name="point_history_point_label">%1$d ポイント</string>
     <string name="dialog_cancel">キャンセル</string>
     <string name="dialog_ok">OK</string>
 </resources>
